libm_dbl: several fixes for a DPFP-based build for the ATSAMD21.

+++ lib/libm_dbl/sqrt.c
@@ -83,7 +83,7 @@ static const double tiny = 1.0e-300;
 double sqrt(double x)
 {
 	double z;
-	int32_t sign = (int)0x80000000;
+	uint32_t sign = 0x80000000;
 	int32_t ix0,s0,q,m,t,i;
 	uint32_t r,t1,s1,ix1,q1;
